backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "openai>=1.68.2",
    "python-dotenv>=1.0.1",
    "sqlmodel>=0.0.24",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
]
